Refine consolidated modules naming and APIs
