+++ index.html
@@ -114,6 +114,6 @@
       </div>
     </div>
 
-    <script type="module" src="/src/main.js"></script>
+    <script type="module" src="/src/uiAndApp.js"></script>
   </body>
 </html>
