--- src/input/controlPrompt.js
@@ -1,355 +1,0 @@
-const STORAGE_KEY = 'grand-theft-scooter:settings';
-
-function detectPreferredTheme() {
-  if (typeof window === 'undefined' || typeof window.matchMedia !== 'function') {
-    return 'dark';
-  }
-  return window.matchMedia('(prefers-color-scheme: light)').matches ? 'light' : 'dark';
-}
-
-function getDefaultSettings() {
-  return {
-    controlScheme: 'wasd',
-    theme: detectPreferredTheme(),
-    cameraSensitivity: 'normal',
-    shadows: false,
-    debugMarkers: false,
-  };
-}
-
-function createUnavailableSettingsManager(reason, { initialWarned = false } = {}) {
-  const fallback = getDefaultSettings();
-  let warned = initialWarned;
-  function warnOnce() {
-    if (warned) return;
-    warned = true;
-    if (typeof console !== 'undefined' && typeof console.warn === 'function') {
-      console.warn(`[settings] ${reason}`);
-    }
-  }
-  return {
-    getControlScheme: () => fallback.controlScheme,
-    getTheme: () => fallback.theme,
-    open: () => {
-      warnOnce();
-    },
-    close: () => {
-      warnOnce();
-    },
-    isOpen: () => false,
-  };
-}
-
-function readStoredSettings() {
-  if (typeof window === 'undefined' || !window.localStorage) {
-    return null;
-  }
-  try {
-    const raw = window.localStorage.getItem(STORAGE_KEY);
-    if (!raw) return null;
-    const parsed = JSON.parse(raw);
-    if (!parsed) return null;
-    const defaults = getDefaultSettings();
-    return {
-      controlScheme: parsed.controlScheme === 'arrows' ? 'arrows' : defaults.controlScheme,
-      theme: parsed.theme === 'light' ? 'light' : 'dark',
-      cameraSensitivity: ['low', 'normal', 'high'].includes(parsed.cameraSensitivity) ? parsed.cameraSensitivity : defaults.cameraSensitivity,
-      shadows: typeof parsed.shadows === 'boolean' ? parsed.shadows : defaults.shadows,
-      debugMarkers: typeof parsed.debugMarkers === 'boolean' ? parsed.debugMarkers : defaults.debugMarkers,
-    };
-  } catch (error) {
-    console.warn('[settings] Unable to parse stored settings, falling back to defaults.', error);
-    return null;
-  }
-}
-
-function persistSettings(settings) {
-  if (typeof window === 'undefined' || !window.localStorage) {
-    return;
-  }
-  try {
-    window.localStorage.setItem(STORAGE_KEY, JSON.stringify(settings));
-  } catch (error) {
-    console.warn('[settings] Failed to persist settings.', error);
-  }
-}
-
-function applyDocumentTheme(theme) {
-  const mode = theme === 'light' ? 'light' : 'dark';
-  document.documentElement.setAttribute('data-theme', mode);
-}
-
-export function createSettingsManager({ onControlSchemeChange, onThemeChange, onCameraSensitivityChange, onGraphicsChange, onDebugChange } = {}) {
-  if (typeof window === 'undefined') {
-    return createUnavailableSettingsManager('Settings unavailable outside a browser context.');
-  }
-
-  const root = document.querySelector('[data-settings]');
-  if (!root) {
-    console.warn('[settings] Settings panel markup missing. Using defaults.');
-    return createUnavailableSettingsManager('Settings panel markup missing. Using defaults.', { initialWarned: true });
-  }
-
-  const controlInputs = Array.from(root.querySelectorAll('[data-control-option]'));
-  const themeInputs = Array.from(root.querySelectorAll('[data-theme-option]'));
-  const sensitivityInputs = Array.from(root.querySelectorAll('[data-sensitivity-option]'));
-  const shadowsInput = root.querySelector('[data-shadows-option]');
-  const debugMarkersInput = root.querySelector('[data-debug-markers-option]');
-
-  const closeTriggers = Array.from(root.querySelectorAll('[data-settings-close]'));
-
-  const settings = {
-    ...getDefaultSettings(),
-    ...readStoredSettings(),
-  };
-
-  let openState = false;
-  let hideTimer = null;
-
-  // Ensure settings are hidden and non-interactive by default
-  root.hidden = true;
-  root.classList.remove('settings--visible');
-  root.style.pointerEvents = 'none';
-
-  function isUiLocked() {
-    try {
-      return !!document.documentElement && document.documentElement.classList.contains('ui-locked');
-    } catch (_) {
-      return false;
-    }
-  }
-
-  function syncControlInputs() {
-    controlInputs.forEach((input) => {
-      input.checked = input.value === settings.controlScheme;
-    });
-  }
-
-  function syncSensitivityInputs() {
-    sensitivityInputs.forEach((input) => {
-      input.checked = input.value === settings.cameraSensitivity;
-    });
-  }
-
-  function syncThemeInputs() {
-    themeInputs.forEach((input) => {
-      input.checked = input.value === settings.theme;
-    });
-  }
-
-  function emitControlChange() {
-    if (typeof onControlSchemeChange === 'function') {
-      onControlSchemeChange(settings.controlScheme);
-    }
-  }
-
-  function emitThemeChange() {
-    if (typeof onThemeChange === 'function') {
-      onThemeChange(settings.theme);
-    }
-  }
-
-  function emitSensitivityChange() {
-    if (typeof onCameraSensitivityChange === 'function') {
-      onCameraSensitivityChange(settings.cameraSensitivity);
-    }
-  }
-
-  function emitGraphicsChange() {
-    if (typeof onGraphicsChange === 'function') {
-      onGraphicsChange({ shadows: !!settings.shadows });
-    }
-  }
-
-  function emitDebugChange() {
-    if (typeof onDebugChange === 'function') {
-      onDebugChange(!!settings.debugMarkers);
-    }
-  }
-
-  function applyControlScheme(nextScheme, { emit = true } = {}) {
-    const normalized = nextScheme === 'arrows' ? 'arrows' : 'wasd';
-    if (settings.controlScheme === normalized) return;
-    settings.controlScheme = normalized;
-    syncControlInputs();
-    persistSettings(settings);
-    if (emit) emitControlChange();
-  }
-
-  function applyCameraSensitivity(next, { emit = true } = {}) {
-    const normalized = ['low', 'normal', 'high'].includes(next) ? next : 'normal';
-    if (settings.cameraSensitivity === normalized) return;
-    settings.cameraSensitivity = normalized;
-    syncSensitivityInputs();
-    persistSettings(settings);
-    if (emit) emitSensitivityChange();
-  }
-
-  function applyTheme(nextTheme, { emit = true } = {}) {
-    const normalized = nextTheme === 'light' ? 'light' : 'dark';
-    if (settings.theme === normalized) return;
-    settings.theme = normalized;
-    applyDocumentTheme(settings.theme);
-    syncThemeInputs();
-    persistSettings(settings);
-    if (emit) emitThemeChange();
-  }
-
-  function open() {
-    if (openState) return;
-    if (isUiLocked()) {
-      return;
-    }
-    openState = true;
-    if (hideTimer) {
-      window.clearTimeout(hideTimer);
-      hideTimer = null;
-    }
-    root.hidden = false;
-    root.style.pointerEvents = 'auto';
-    requestAnimationFrame(() => {
-      root.classList.add('settings--visible');
-    });
-    const focusTarget = root.querySelector('input:checked') || root.querySelector('input');
-    if (focusTarget) {
-      focusTarget.focus({ preventScroll: true });
-    }
-  }
-
-  function close() {
-    if (!openState) return;
-    openState = false;
-    root.classList.remove('settings--visible');
-    root.style.pointerEvents = 'none';
-    hideTimer = window.setTimeout(() => {
-      if (!openState) {
-        root.hidden = true;
-      }
-    }, 200);
-  }
-
-  function isOpen() {
-    return openState;
-  }
-
-  function handleKeydown(event) {
-    if (event.key !== 'Escape') return;
-    if (!openState && isUiLocked()) {
-      event.preventDefault();
-      return;
-
-    }
-    if (openState) {
-      event.preventDefault();
-      close();
-    } else {
-      open();
-    }
-  }
-
-  function handleTransitionEnd(event) {
-    if (event.target !== root || event.propertyName !== 'opacity') return;
-    if (!openState) {
-      root.hidden = true;
-    }
-  }
-
-  window.addEventListener('keydown', handleKeydown);
-  root.addEventListener('transitionend', handleTransitionEnd);
-
-  controlInputs.forEach((input) => {
-    input.addEventListener('change', () => {
-      if (!openState) return;
-      applyControlScheme(input.value);
-    });
-  });
-
-  themeInputs.forEach((input) => {
-    input.addEventListener('change', () => {
-      if (!openState || isUiLocked()) return;
-      applyTheme(input.value);
-    });
-  });
-
-  closeTriggers.forEach((element) => {
-    element.addEventListener('click', (event) => {
-      event.preventDefault();
-      close();
-    });
-
-  });
-
-  // Initial sync applies the saved state, then notifies the rest of the app once.
-  syncControlInputs();
-  applyDocumentTheme(settings.theme);
-  syncThemeInputs();
-  syncSensitivityInputs();
-
-  sensitivityInputs.forEach((input) => {
-    input.addEventListener('change', () => {
-      if (!openState) return;
-      applyCameraSensitivity(input.value);
-    });
-  });
-
-  function syncShadowsInput() {
-    if (shadowsInput) {
-      shadowsInput.checked = !!settings.shadows;
-    }
-  }
-
-  function syncDebugMarkersInput() {
-    if (debugMarkersInput) {
-      debugMarkersInput.checked = !!settings.debugMarkers;
-    }
-  }
-
-  if (shadowsInput) {
-    shadowsInput.addEventListener('change', () => {
-      if (!openState || isUiLocked()) { // block background changes while closed/locked
-        shadowsInput.checked = !!settings.shadows; // snap back UI state
-        return;
-      }
-      settings.shadows = !!shadowsInput.checked;
-      persistSettings(settings);
-      emitGraphicsChange();
-    });
-  }
-
-  if (debugMarkersInput) {
-    debugMarkersInput.addEventListener('change', () => {
-      if (!openState || isUiLocked()) {
-        debugMarkersInput.checked = !!settings.debugMarkers;
-        return;
-      }
-      settings.debugMarkers = !!debugMarkersInput.checked;
-      persistSettings(settings);
-      emitDebugChange();
-    });
-  }
-
-  // Initial sync applies the saved state, then notifies the rest of the app once.
-  syncControlInputs();
-  applyDocumentTheme(settings.theme);
-  syncThemeInputs();
-  syncSensitivityInputs();
-  syncShadowsInput();
-  syncDebugMarkersInput();
-
-  emitSensitivityChange();
-  emitControlChange();
-  emitThemeChange();
-  emitGraphicsChange();
-  emitDebugChange();
-
-  return {
-    getControlScheme: () => settings.controlScheme,
-    getTheme: () => settings.theme,
-    getCameraSensitivity: () => settings.cameraSensitivity,
-    getShadowsEnabled: () => !!settings.shadows,
-    getDebugMarkersEnabled: () => !!settings.debugMarkers,
-    open,
-    close,
-    isOpen,
-  };
-}
